# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for HR Platform with Bulgarian payroll",
        "title": "HR Platform API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/payslips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "List payslips for a specific payroll run",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payslips"
                ],
                "summary": "List payslips",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payroll run ID (UUID)",
                        "name": "payroll_run_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payslips/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Retrieve a single payslip by UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payslips"
                ],
                "summary": "Get payslip by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payslip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payslips/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ],
                "description": "Confirm a payslip has been reviewed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payslips"
                ],
                "summary": "Confirm a payslip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payslip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handler.ErrorBody"
                }
            }
        },
        "handler.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}